requirements.txt
# Core Framework
flask==3.1.3
flask-cors==6.0.2
waitress==3.0.2

# Database (with pre-compiled PostgreSQL binaries)
psycopg[binary]==3.3.3
psycopg-pool==3.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2
